requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pypdf>=3.0.0
tabula-py>=2.3.0
opencv-python>=4.5.0
Pillow>=9.0.0
pytesseract>=0.3.0
pdf2image>=1.16.0
torch>=2.0.0
transformers>=4.30.0
jupyter>=1.0.0
notebook>=6.0.0